+++ src/components/contexts/UserContext.tsx
@@ -42,6 +42,8 @@ export const UserProvider = ({ children }: { children: ReactNode }): ReactElemen
     cardFarming: {
       listGames: true,
       allGames: false,
+      nextTaskCheckbox: false,
+      nextTask: null,
       credentials: null,
       userSummary: null,
       gamesWithDrops: 0,
@@ -50,6 +52,8 @@ export const UserProvider = ({ children }: { children: ReactNode }): ReactElemen
     achievementUnlocker: {
       idle: true,
       hidden: false,
+      nextTaskCheckbox: false,
+      nextTask: null,
       schedule: false,
       scheduleFrom: new Time(8, 30, 0, 0),
       scheduleTo: new Time(23, 0, 0, 0),