+++ src/components/settings/AchievementSettings.tsx
@@ -1,11 +1,12 @@
 import type { ReactElement } from 'react'
 
-import { cn, Slider, TimeInput } from '@heroui/react'
+import { cn, Select, SelectItem, Slider, TimeInput } from '@heroui/react'
 import { useTranslation } from 'react-i18next'
 
 import { useUserContext } from '@/components/contexts/UserContext'
 import SettingsCheckbox from '@/components/settings/SettingsCheckbox'
 import {
+  handleNextTaskChange,
   handleScheduleChange,
   handleSliderChange,
   useAchievementSettings,
@@ -16,6 +17,17 @@ export default function AchievementSettings(): ReactElement {
   const { userSummary, userSettings, setUserSettings } = useUserContext()
   const { sliderLabel, setSliderLabel } = useAchievementSettings()
 
+  const taskOptions = [
+    {
+      key: 'cardFarming',
+      label: t('common.cardFarming'),
+    },
+    {
+      key: 'autoIdle',
+      label: t('customLists.autoIdle.title'),
+    },
+  ]
+
   return (
     <div className='relative flex flex-col gap-4'>
       <div className='flex flex-col gap-4 border border-border rounded-lg p-3 bg-titlebar'>
@@ -24,6 +36,48 @@ export default function AchievementSettings(): ReactElement {
         <SettingsCheckbox type='achievementUnlocker' name='idle' content={t('settings.achievementUnlocker.idle')} />
 
         <SettingsCheckbox type='achievementUnlocker' name='hidden' content={t('settings.achievementUnlocker.hidden')} />
+
+        {userSettings.general.useBeta && (
+          <div className='flex items-center gap-2'>
+            <SettingsCheckbox type='achievementUnlocker' name='nextTaskCheckbox' content={t('common.nextTask')} />
+
+            <Select
+              size='sm'
+              aria-label='nextTask'
+              disallowEmptySelection
+              radius='none'
+              items={taskOptions}
+              className='w-[200px]'
+              placeholder='Select an option'
+              classNames={{
+                listbox: ['p-0'],
+                value: ['text-sm !text-content'],
+                trigger: cn(
+                  'bg-input border border-border data-[hover=true]:!bg-inputhover',
+                  'data-[open=true]:!bg-input duration-100 rounded-lg',
+                ),
+                popoverContent: ['bg-titlebar border border-border rounded-lg justify-start !text-content'],
+              }}
+              isDisabled={!userSettings.achievementUnlocker.nextTaskCheckbox}
+              defaultSelectedKeys={
+                userSettings.achievementUnlocker.nextTask ? [userSettings.achievementUnlocker.nextTask] : []
+              }
+              onSelectionChange={e => {
+                handleNextTaskChange(e.currentKey!, userSummary, setUserSettings)
+              }}
+            >
+              {item => (
+                <SelectItem
+                  classNames={{
+                    base: ['data-[hover=true]:!bg-titlehover data-[hover=true]:!text-content'],
+                  }}
+                >
+                  {item.label}
+                </SelectItem>
+              )}
+            </Select>
+          </div>
+        )}
 
         <div className='flex items-center gap-2'>
           <SettingsCheckbox