+++ src/hooks/settings/useCardSettings.ts
@@ -326,3 +326,17 @@ export const handleClear = async (
     logEvent(`[Error] in (handleClear): ${error}`)
   }
 }
+
+export const handleNextTaskChange = async (
+  currentKey: string,
+  userSummary: UserSummary,
+  setUserSettings: Dispatch<SetStateAction<UserSettings>>,
+): Promise<void> => {
+  const response = await invoke<InvokeSettings>('update_user_settings', {
+    steamId: userSummary?.steamId,
+    key: 'cardFarming.nextTask',
+    value: currentKey,
+  })
+
+  setUserSettings(response.settings)
+}